Move auto_roll_logger and filename out of DB
Summary: It is confusing to have auto_roll_logger to stay under DB, which has nothing to do with database. Move filename together as it is a dependency.

Test Plan: Run all unit tests
