+++ db/builder.cc
@@ -16,7 +16,6 @@
 #include "db/compaction_iterator.h"
 #include "db/dbformat.h"
 #include "db/event_helpers.h"
-#include "db/filename.h"
 #include "db/internal_stats.h"
 #include "db/merge_helper.h"
 #include "db/table_cache.h"
@@ -29,6 +28,7 @@
 #include "table/block_based_table_builder.h"
 #include "table/internal_iterator.h"
 #include "util/file_reader_writer.h"
+#include "util/filename.h"
 #include "util/iostats_context_imp.h"
 #include "util/stop_watch.h"
 #include "util/thread_status_util.h"
