+++ db/corruption_test.cc
@@ -11,19 +11,19 @@
 
 #include "rocksdb/db.h"
 
-#include <inttypes.h>
 #include <errno.h>
 #include <fcntl.h>
+#include <inttypes.h>
 #include <sys/stat.h>
 #include <sys/types.h>
+#include "db/db_impl.h"
+#include "db/log_format.h"
+#include "db/version_set.h"
 #include "rocksdb/cache.h"
 #include "rocksdb/env.h"
 #include "rocksdb/table.h"
 #include "rocksdb/write_batch.h"
-#include "db/db_impl.h"
-#include "db/filename.h"
-#include "db/log_format.h"
-#include "db/version_set.h"
+#include "util/filename.h"
 #include "util/logging.h"
 #include "util/testharness.h"
 #include "util/testutil.h"
