+++ db/deletefile_test.cc
@@ -14,12 +14,12 @@
 #include <string>
 #include <vector>
 #include "db/db_impl.h"
-#include "db/filename.h"
 #include "db/version_set.h"
 #include "db/write_batch_internal.h"
 #include "rocksdb/db.h"
 #include "rocksdb/env.h"
 #include "rocksdb/transaction_log.h"
+#include "util/filename.h"
 #include "util/string_util.h"
 #include "util/sync_point.h"
 #include "util/testharness.h"
