+++ db/filename_test.cc
@@ -7,7 +7,7 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file. See the AUTHORS file for names of contributors.
 
-#include "db/filename.h"
+#include "util/filename.h"
 
 #include "db/dbformat.h"
 #include "port/port.h"
