+++ db/flush_job.cc
@@ -22,7 +22,6 @@
 #include "db/db_iter.h"
 #include "db/dbformat.h"
 #include "db/event_helpers.h"
-#include "db/filename.h"
 #include "db/log_reader.h"
 #include "db/log_writer.h"
 #include "db/memtable.h"
@@ -44,6 +43,7 @@
 #include "util/coding.h"
 #include "util/event_logger.h"
 #include "util/file_util.h"
+#include "util/filename.h"
 #include "util/iostats_context_imp.h"
 #include "util/log_buffer.h"
 #include "util/logging.h"
