+++ db/repair.cc
@@ -68,7 +68,6 @@
 #include "db/builder.h"
 #include "db/db_impl.h"
 #include "db/dbformat.h"
-#include "db/filename.h"
 #include "db/log_reader.h"
 #include "db/log_writer.h"
 #include "db/memtable.h"
@@ -83,6 +82,7 @@
 #include "table/scoped_arena_iterator.h"
 #include "util/cf_options.h"
 #include "util/file_reader_writer.h"
+#include "util/filename.h"
 #include "util/string_util.h"
 
 namespace rocksdb {
