+++ db/transaction_log_impl.h
@@ -7,7 +7,6 @@
 #ifndef ROCKSDB_LITE
 #include <vector>
 
-#include "db/filename.h"
 #include "db/log_reader.h"
 #include "db/version_set.h"
 #include "port/port.h"
@@ -16,6 +15,7 @@
 #include "rocksdb/transaction_log.h"
 #include "rocksdb/types.h"
 #include "util/db_options.h"
+#include "util/filename.h"
 
 namespace rocksdb {
