+++ db/wal_manager.cc
@@ -18,10 +18,9 @@
 #include <vector>
 #include <memory>
 
-#include "db/filename.h"
-#include "db/transaction_log_impl.h"
 #include "db/log_reader.h"
 #include "db/log_writer.h"
+#include "db/transaction_log_impl.h"
 #include "db/write_batch_internal.h"
 #include "port/port.h"
 #include "rocksdb/env.h"
@@ -29,10 +28,11 @@
 #include "rocksdb/write_batch.h"
 #include "util/coding.h"
 #include "util/file_reader_writer.h"
+#include "util/filename.h"
 #include "util/logging.h"
 #include "util/mutexlock.h"
-#include "util/sync_point.h"
 #include "util/string_util.h"
+#include "util/sync_point.h"
 
 namespace rocksdb {
