+++ tools/rdb/db_wrapper.cc
@@ -4,10 +4,10 @@
 #include <v8.h>
 #include <node.h>
 
-#include "db_wrapper.h"
+#include "db/_wrapper.h"
 #include "rocksdb/db.h"
-#include "rocksdb/slice.h"
 #include "rocksdb/options.h"
+#include "rocksdb/slice.h"
 
 namespace {
   void printWithBackSlashes(std::string str) {
