+++ tools/rdb/rdb.cc
@@ -2,9 +2,9 @@
 #define BUILDING_NODE_EXTENSION
 #endif
 
-#include <v8.h>
 #include <node.h>
-#include "db_wrapper.h"
+#include <v8.h>
+#include "db/_wrapper.h"
 
 using namespace v8;
