+++ util/auto_roll_logger.cc
@@ -3,7 +3,7 @@
 //  LICENSE file in the root directory of this source tree. An additional grant
 //  of patent rights can be found in the PATENTS file in the same directory.
 //
-#include "db/auto_roll_logger.h"
+#include "util/auto_roll_logger.h"
 #include "util/mutexlock.h"
 
 namespace rocksdb {
