+++ util/auto_roll_logger.h
@@ -10,11 +10,11 @@
 #include <list>
 #include <string>
 
-#include "db/filename.h"
 #include "port/port.h"
 #include "port/util_logger.h"
-#include "util/sync_point.h"
+#include "util/filename.h"
 #include "util/mutexlock.h"
+#include "util/sync_point.h"
 
 namespace rocksdb {
