+++ util/fault_injection_test_env.h
@@ -18,10 +18,10 @@
 #include <set>
 #include <string>
 
-#include "db/filename.h"
 #include "db/version_set.h"
 #include "rocksdb/db.h"
 #include "rocksdb/env.h"
+#include "util/filename.h"
 #include "util/mock_env.h"
 #include "util/mutexlock.h"
