+++ utilities/backupable/backupable_db_test.cc
@@ -13,7 +13,6 @@
 #include <string>
 
 #include "db/db_impl.h"
-#include "db/filename.h"
 #include "port/port.h"
 #include "port/stack_trace.h"
 #include "rocksdb/rate_limiter.h"
@@ -23,6 +22,7 @@
 #include "rocksdb/utilities/options_util.h"
 #include "util/env_chroot.h"
 #include "util/file_reader_writer.h"
+#include "util/filename.h"
 #include "util/mutexlock.h"
 #include "util/random.h"
 #include "util/stderr_logger.h"
