+++ utilities/checkpoint/checkpoint.cc
@@ -18,14 +18,14 @@
 #include <inttypes.h>
 #include <algorithm>
 #include <string>
-#include "db/filename.h"
 #include "db/wal_manager.h"
+#include "port/port.h"
 #include "rocksdb/db.h"
 #include "rocksdb/env.h"
 #include "rocksdb/transaction_log.h"
 #include "util/file_util.h"
+#include "util/filename.h"
 #include "util/sync_point.h"
-#include "port/port.h"
 
 namespace rocksdb {
