+++ utilities/convenience/info_log_finder.cc
@@ -7,9 +7,9 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-#include "db/filename.h"
-#include "rocksdb/env.h"
 #include "rocksdb/utilities/info_log_finder.h"
+#include "rocksdb/env.h"
+#include "util/filename.h"
 
 namespace rocksdb {
