+++ utilities/date_tiered/date_tiered_db_impl.cc
@@ -9,7 +9,6 @@
 
 #include "db/db_impl.h"
 #include "db/db_iter.h"
-#include "db/filename.h"
 #include "db/write_batch_internal.h"
 #include "rocksdb/convenience.h"
 #include "rocksdb/env.h"
@@ -17,6 +16,7 @@
 #include "rocksdb/utilities/date_tiered_db.h"
 #include "table/merging_iterator.h"
 #include "util/coding.h"
+#include "util/filename.h"
 #include "util/instrumented_mutex.h"
 #include "util/options_helper.h"
 #include "util/string_util.h"
