+++ utilities/geodb/geodb_impl.cc
@@ -11,14 +11,13 @@
 #define __STDC_FORMAT_MACROS
 #endif
 
-#include <vector>
+#include <limits>
 #include <map>
 #include <string>
-#include <limits>
-#include "db/filename.h"
+#include <vector>
 #include "util/coding.h"
+#include "util/filename.h"
 #include "util/string_util.h"
-
 
 //
 // There are two types of keys. The first type of key-values
