+++ utilities/options/options_util.cc
@@ -7,8 +7,8 @@
 
 #include "rocksdb/utilities/options_util.h"
 
-#include "db/filename.h"
 #include "rocksdb/options.h"
+#include "util/filename.h"
 #include "util/options_parser.h"
 
 namespace rocksdb {
